requirements.txt
uvicorn==0.22.0
fastapi==0.95.2
Jinja2==3.1.2
sqlalchemy==2.0.22
pydantic==1.10.7
PyYAML==6.0
reportlab==4.4.4
python-multipart==0.0.6
pytest==7.4.0
requests==2.31.0
python-dotenv==1.0.0
httpx==0.24.1
